clean validation architecture
