+++ includes/config_parser/ValidationStructureConfig.hpp
@@ -3,10 +3,13 @@
 
 #include <string>
 
-void isEmptyBraceOrSemicolonLine(const std::string &trimmedLine, int &lineCont, const std::string &filePath);
-
 void checkEmptyBraceOrSemicolon(const std::string &trimmedLine, int lineCont, const std::string &filePath,
                                 std::vector<std::string> &errors);
+
+void checkInvalidCharacters(const std::string &trimmedLine, int lineCont, std::vector<std::string> &errors);
+
+void checkBraceBalance(int contOpenKey, int contCloseKey, int firstOpenKey, int lastCloseKey, 
+                    const std::string &filePath, std::vector<std::string> &errors);
 
 int contOpenKeys(const std::string &trimmedLine, int &lineCont, int &contOpenKey);
 
@@ -15,13 +18,6 @@ int contCloseKeys(const std::string &trimmedLine, int &lineCont, int &contCloseK
 void processConfigLine(const std::string &trimmedLine, int &lineCont, int &contOpenKey,
                     int &contCloseKey, int &firstOpenKey, int &lastCloseKey);
 
-void resultProcesConfigLine(int contOpenKey, int contCloseKey, int firstOpenKey,
-                            int lastCloseKey, const std::string &filePath);
-
-void firstNonAlNumChar(const std::string &trimmedLine, int &lineCont, const std::string &filePath);
-
 bool validateStructure(const std::string &filePath, std::vector<std::string> &errors);
-
-void validationStructureConfigFile(const std::string &filePath);
 
 #endif
